Smart Resume & Job Fit Analyzer/backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic<2.0.0
python-multipart>=0.0.6

# Resume Parsing
pdfplumber>=0.10.3
python-docx>=1.1.0

# NLP & Matching
spacy>=3.7.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
rapidfuzz>=3.5.0
scikit-learn>=1.4.0

# Configuration
pyyaml>=6.0.1

# PDF Export
reportlab>=4.0.0

# Testing
pytest>=7.4.0
httpx>=0.26.0
pytest-cov>=4.1.0
